add visualize_libary
split misc_libary into visalize_libary
